move Expected type into detail/expected_impl.hpp
